# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'QMS API Documentation'
  description: 'Secure messenger API with end-to-end encryption'
  contact:
    name: Heruvim6w
    email: heruvim.6w@gmail.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'QMS API Server'
  -
    url: 'http://localhost:8000'
    description: 'Development Server'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
paths:
...
  /api/v1/users/profile:
    get:
      tags:
        - Users
      summary: 'Get current user profile'
      description: 'Retrieve the complete profile of the authenticated user including personal data, UIN, username, status, and created date.'
      operationId: getProfile
      responses:
        200:
          description: 'User profile data'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        401:
          description: Unauthenticated
      security:
        -
          bearerAuth: []
  /api/v1/users/username:
    post:
      tags:
        - Users
      summary: 'Set or update custom username'
      description: 'Create or change a custom unique username (like in Telegram). Username must be 3-20 characters, Latin letters, digits, underscore, and dash only. Can be changed anytime. UIN remains unchanged.'
      operationId: setUsername
      requestBody:
        description: 'New username'
        required: true
        content:
          application/json:
            schema:
              required:
                - username
              properties:
                username:
                  type: string
                  maxLength: 20
                  minLength: 3
                  pattern: '^[a-zA-Z0-9_-]+$'
                  example: john_doe
              type: object
      responses:
        200:
          description: 'Username updated successfully'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  username:
                    type: string
                    example: john_doe
                type: object
        400:
          description: 'Invalid username format'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Username must contain 3-20 characters (letters, digits, underscore, dash)'
                type: object
        422:
          description: 'Validation error or username already taken'
      security:
        -
          bearerAuth: []
  /api/v1/users/search:
    get:
      tags:
        - Users
      summary: 'Search user by UIN or username'
      description: 'Search for a user by their 8-digit UIN or custom username. Can be used to find users for adding to contacts or group chats. Returns basic user profile information.'
      operationId: searchUser
      parameters:
        -
          name: query
          in: query
          description: 'Search query - either UIN (8 digits) or username (3-20 chars, latin only)'
          required: true
          schema:
            type: string
            minLength: 3
            example: '12345678'
      responses:
        200:
          description: 'User found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchUserResponse'
        400:
          description: 'Invalid search query format'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Invalid search query'
                type: object
        404:
          description: 'User not found'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'User not found'
                type: object
      security:
        -
          bearerAuth: []
  '/api/v1/users/{identifier}':
    get:
      tags:
        - Users
      summary: 'Get user profile by UIN or username'
      description: 'Retrieve detailed user profile information using their UIN (8-digit identifier) or custom username. Returns public profile information including status and last seen time.'
      operationId: getUserByIdentifier
      parameters:
        -
          name: identifier
          in: path
          description: 'User identifier - either UIN (8 digits) or username (3-20 chars)'
          required: true
          schema:
            type: string
            example: john_doe
      responses:
        200:
          description: 'User profile data'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchUserResponse'
        404:
          description: 'User not found with the given identifier'
      security:
        -
          bearerAuth: []
  /api/v1/sessions:
    get:
      tags:
        - Users
      summary: 'Get all active sessions for current user'
      description: 'Retrieve a list of all currently active sessions (confirmed login devices). Each session contains device info, confirmation date, and expiration date. Helps user track and manage their connected devices.'
      operationId: getSessions
      responses:
        200:
          description: 'List of active sessions across all devices'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    id:
                      description: 'Session ID'
                      type: integer
                    device_name:
                      description: 'User-friendly device name'
                      type: string
                      example: 'iPhone 13'
                    ip_address:
                      description: 'IP address where device logged in'
                      type: string
                      format: ipv4
                      example: 192.168.1.1
                    confirmed_at:
                      description: 'When this device was confirmed'
                      type: string
                      format: date-time
                    expires_at:
                      description: 'When this session expires'
                      type: string
                      format: date-time
                  type: object
      security:
        -
          bearerAuth: []
  '/api/v1/sessions/{sessionId}':
    delete:
      tags:
        - Users
      summary: 'End a specific session (logout from device)'
      description: 'Terminate a specific session and logout from that device. Other active sessions remain unaffected. The session will no longer be able to use their JWT tokens.'
      operationId: endSession
      parameters:
        -
          name: sessionId
          in: path
          description: 'Unique session identifier to terminate'
          required: true
          schema:
            type: integer
            example: 42
      responses:
        200:
          description: 'Session successfully terminated'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 'Session ended'
                type: object
        404:
          description: 'Session not found or does not belong to current user'
      security:
        -
          bearerAuth: []
  /api/v1/users/status:
    post:
      tags:
        - Users
      summary: 'Set user online status and custom status text'
      description: "Update the user's online status (mood) and optional custom status text. User can only select offline status indirectly (via inactivity). Custom status supports emojis and has a max of 50 characters. Status is displayed to friends in their contact list."
      operationId: setStatus
      requestBody:
        description: 'Status information to set'
        required: true
        content:
          application/json:
            schema:
              required:
                - online_status
              properties:
                online_status:
                  description: 'Selected mood/status'
                  type: string
                  enum:
                    - online
                    - chatty
                    - angry
                    - depressed
                    - home
                    - work
                    - eating
                    - away
                    - unavailable
                    - busy
                    - do_not_disturb
                  example: chatty
                custom_status:
                  description: 'Optional custom status text (max 50 characters, supports emoji and 😀)'
                  type: string
                  maxLength: 50
                  example: 'На встрече 🎯'
                  nullable: true
              type: object
      responses:
        200:
          description: 'Status updated successfully'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  online_status:
                    description: 'Set online status'
                    type: string
                    example: chatty
                  display_status:
                    description: 'Combined display string with localization'
                    type: string
                    example: 'Готов поболтать - На встрече 🎯'
                type: object
        422:
          description: 'Validation error (invalid status, custom status too long, etc)'
      security:
        -
          bearerAuth: []
  /api/v1/users/status/available:
    get:
      tags:
        - Users
      summary: 'Get list of available online statuses'
      description: 'Retrieve list of all available online statuses (moods) that user can select from. Includes both localized display names and status keys for client-side usage.'
      operationId: getAvailableStatuses
      responses:
        200:
          description: 'List of available online statuses with localized names'
          content:
            application/json:
              schema:
                properties:
                  statuses:
                    description: 'Key-value pairs of status key and localized display name'
                    type: object
                    additionalProperties:
                      type: string
                      property: null
                type: object
      security:
        -
          bearerAuth: []
  '/api/v1/users/{identifier}/status':
    get:
      tags:
        - Users
      summary: 'Get user status (for friends/contacts list)'
      description: 'Retrieve the current status of a user, including whether they are online and their selected status mood. If offline, shows time of last seen. Used for displaying user status in contacts list.'
      operationId: getUserStatus
      parameters:
        -
          name: identifier
          in: path
          description: 'User UIN (8 digits) or username'
          required: true
          schema:
            type: string
            example: john_doe
      responses:
        200:
          description: 'User status information'
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  uin:
                    type: string
                  is_online:
                    description: 'true if user is online, false if offline'
                    type: boolean
                  status:
                    description: 'Localized status name'
                    type: string
                    example: 'Готов поболтать'
                  status_key:
                    description: 'Status key for client-side use'
                    type: string
                    example: chatty
                  last_seen:
                    description: 'Human-readable last seen time (only if offline)'
                    type: string
                    example: '3 hours ago'
                    nullable: true
                type: object
        404:
          description: 'User not found'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    UserProfileResponse:
      title: 'User Profile'
      description: 'Complete user profile information'
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        uin:
          description: '8-digit unique identifier'
          type: string
        username:
          description: 'Custom username (3-20 chars, latin only)'
          type: string
          nullable: true
        status:
          description: 'Current online status'
          type: string
          enum:
            - online
            - offline
        online_status:
          description: 'Selected status mood'
          type: string
          enum:
            - online
            - chatty
            - angry
            - depressed
            - home
            - work
            - eating
            - away
            - unavailable
            - busy
            - do_not_disturb
        custom_status:
          description: 'User custom status text with emoji support'
          type: string
          maxLength: 50
          nullable: true
        last_seen_at:
          description: 'Last online timestamp'
          type: string
          format: date-time
          nullable: true
        locale:
          description: 'User interface language'
          type: string
          example: ru
        created_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []